# refactor: optimize `Path.Relative.append`