The previous implementation would create an intermediate string after `x ^ "/"` that would be immediately be thrown away. With the new implementation, no temporary garbage is created.

Should be compared to https://github.com/ocaml/dune/pull/10094

```
┌────────────────────────────────────────────────────────────────────────────┬────────────┬───────────┬──────────┬──────────┬────────────┐
│ Name                                                                       │   Time/Run │   mWd/Run │ mjWd/Run │ Prom/Run │ Percentage │
├────────────────────────────────────────────────────────────────────────────┼────────────┼───────────┼──────────┼──────────┼────────────┤
│ [bench/micro/path_bench.ml] Path.Local.append:left root                    │    15.54ns │           │          │          │      1.28% │
│ [bench/micro/path_bench.ml] Path.Local.append:right root                   │    15.70ns │           │          │          │      1.30% │
│ [bench/micro/path_bench.ml] Path.Local.append:short paths                  │    38.01ns │     3.00w │          │          │      3.14% │
│ [bench/micro/path_bench.ml] Path.Local.append:long paths                   │    47.01ns │    61.00w │          │          │      3.88% │
└────────────────────────────────────────────────────────────────────────────┴────────────┴───────────┴──────────┴──────────┴────────────┘
```

From the larger benchmarks:
* Promoted words are decreased by up to 0.44%
* Minor collections are reduced by 1%
* The monorepo benchmark shows a considerable reduction in zero builds, but that seems highly unlikely.

Overall, the optimization itself is rather simple so I think it should be considered.